Move spec service under support/
